% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{add_font}
\alias{add_font}
\title{Download google font}
\usage{
add_font(
  path_data,
  font_name = "Lora",
  font_url = "https://fonts.google.com/download?family=Lora"
)
}
\arguments{
\item{font_url}{}
}
\description{
Download google font
}
