% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_CV_a2.R
\name{calc_CV_a2}
\alias{calc_CV_a2}
\title{Calculate CV based on CV model (approach 2)}
\usage{
calc_CV_a2(n, d, a1, a2, cv_params)
}
\arguments{
\item{cv_params}{}
}
\description{
Calculate CV based on CV model (approach 2)
}
