% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_time.R
\name{calc_time}
\alias{calc_time}
\title{Calculate unit plot time based on plot design}
\usage{
calc_time(
  unit_times,
  plot_design,
  nest_design,
  progress_id = NULL,
  session = NULL
)
}
\arguments{
\item{session}{}
}
\description{
Calculate unit plot time based on plot design
}
