% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_avitabile.R
\name{get_avitabile}
\alias{get_avitabile}
\title{Download Avitabile raster biomass data}
\usage{
get_avitabile(
  path_data,
  progress_id = NULL,
  session = NULL,
  sf_aoi = NULL,
  url =
    "http://lucid.wur.nl/storage/downloads/high-carbon-ecosystems/Avitabile_AGB_Map.zip"
)
}
\arguments{
\item{url}{}
}
\description{
Download Avitabile raster biomass data
}
