% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_santoro.R
\name{get_santoro_tiles}
\alias{get_santoro_tiles}
\title{Get tile names for downloading Santoro 2018 biomass raster data}
\usage{
get_santoro_tiles(sf_aoi)
}
\arguments{
\item{sf_aoi}{}
}
\description{
Get tile names for downloading Santoro 2018 biomass raster data
}
