% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_params_server.R
\name{mod_params_server}
\alias{mod_params_server}
\title{Optimization parameters module server function}
\usage{
mod_params_server(id, rv)
}
\arguments{
\item{rv}{}
}
\description{
Optimization parameters module server function
}
