% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_params_UI.R
\name{mod_params_UI}
\alias{mod_params_UI}
\title{Optimization parameters module UI function}
\usage{
mod_params_UI(id)
}
\arguments{
\item{id}{}
}
\description{
Optimization parameters module UI function
}
