% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_results_server.R
\name{mod_results_server}
\alias{mod_results_server}
\title{Optimization results module server function}
\usage{
mod_results_server(id, rv)
}
\arguments{
\item{rv}{}
}
\description{
Optimization results module server function
}
