% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_results_UI.R
\name{mod_results_UI}
\alias{mod_results_UI}
\title{Optimization results module server function}
\usage{
mod_results_UI(id)
}
\arguments{
\item{id}{}
}
\description{
Optimization results module server function
}
