% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_time_server.R
\name{mod_time_server}
\alias{mod_time_server}
\title{Inventory unit time module server function}
\usage{
mod_time_server(id, rv)
}
\arguments{
\item{rv}{}
}
\description{
Inventory unit time module server function
}
