% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submod_CV_a1_server.R
\name{submod_CV_a1_server}
\alias{submod_CV_a1_server}
\title{CV approach 1 submodule server function}
\usage{
submod_CV_a1_server(id, rv, rv_cv)
}
\arguments{
\item{rv_cv}{}
}
\description{
CV approach 1 submodule server function
}
