% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submod_CV_a2_server.R
\name{submod_CV_a2_server}
\alias{submod_CV_a2_server}
\title{CV approach 2 submodule server function}
\usage{
submod_CV_a2_server(id, rv)
}
\arguments{
\item{rv}{}
}
\description{
CV approach 2 submodule server function
}
