% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submod_CV_intro_server.R
\name{submod_CV_intro_server}
\alias{submod_CV_intro_server}
\title{CV introduction submodule server function}
\usage{
submod_CV_intro_server(id, rv)
}
\arguments{
\item{rv}{}
}
\description{
CV introduction submodule server function
}
