% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_CV_a1.R
\name{calc_CV_a1}
\alias{calc_CV_a1}
\title{Calculate CV based on plot design and initial CV for approach 1}
\usage{
calc_CV_a1(cv_init, area_init, area_opti)
}
\arguments{
\item{area_opti}{}
}
\description{
Calculate CV based on plot design and initial CV for approach 1
}
