% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_avitabile.R
\name{get_avitabile}
\alias{get_avitabile}
\title{Download Avitabile raster biomass data}
\usage{
get_avitabile(
  path_data,
  progress_id = NULL,
  session = NULL,
  sf_aoi = NULL,
  url =
    "http://lucid.wur.nl/storage/downloads/high-carbon-ecosystems/Avitabile_AGB_Map.zip"
)
}
\arguments{
\item{path_data}{a path to download or load the biomass map.}

\item{progress_id, session}{in a Shiny context, links to a shinyWidgets::progressBar() object}

\item{sf_aoi}{a simple feature object to clip the raster data}

\item{url}{url for downloading the map. Defaults to:
http://lucid.wur.nl/storage/downloads/high-carbon-ecosystems/Avitabile_AGB_Map.zip}
}
\description{
This function makes the biomass map from Avitabile et al. 2016
(An integrated pan-tropical biomass map using multiple reference
datasets) available to use in R. It requires a directory path, checks
if the data is already there, if not download the data, and load in
the R environement as a terra::rast object. Additionally if the user
provide a spatial area boundaries (as simple feature object, see sf)
the raster biomass data is clipped to the sf object extent.
}
\examples{
path_data <- tempdir(dir.create("data", showWarnings = F))

rs_avitabile <- get_avitabile(path_data = path_data)
plot(rs_avitabile)

unlink(path_data)

}
