% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_santoro.R
\name{get_santoro}
\alias{get_santoro}
\title{Download Santoro 2018 tiles based on users' area of interest}
\usage{
get_santoro(
  path_data,
  progress_id = NULL,
  session = NULL,
  sf_aoi = NULL,
  tile_name = "N00E140_agb.zip",
  url = "globbiomass.org/wp-content/uploads/GB_Maps/"
)
}
\arguments{
\item{url}{}
}
\description{
Download Santoro 2018 tiles based on users' area of interest
}
